feat(dashboard):product table page

+++ Server/database/migrations/2024_05_22_094522_create_carts_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('carts', function (Blueprint $table) {
             $table->id();
-            $table->integer('user_id')->nullable();
+            $table->unsignedBigInteger('user_id')->nullable();
             $table->integer('total_price')->default(1);
             $table->timestamps();
